refactor: decompose ini draw analysis
